feat(boot): implement boot animation feature
Added a new boot animation application to enhance the user experience on startup. This includes the addition of `BootAnimationApp` class, which manages the display of a series of frames as an animated sequence. The application is set as the default active app during setup. The update involves changes to several files to support this new feature, such as including necessary headers and updating image resources for the animation frames.

Sequence is tunable with frameInterval and sequenceDuration. May be more desirable to keep the image assets, if any, in the BootAnimation folder nearby (makes things easier for the linker to find).